requirements.txt
asgiref==3.2.7
certifi==2020.4.5.1
chardet==3.0.4
curlify==2.2.1
Django==3.0.5
djangorestframework==3.11.0
facebook-business==6.0.3
gunicorn==20.0.4
idna==2.9
pycountry==19.8.18
pytz==2019.3
requests==2.23.0
six==1.14.0
sqlparse==0.3.1
urllib3==1.25.9
whitenoise==5.0.1